+++ src/main/java/taskList/tasks/Event.java
@@ -1,4 +1,4 @@
-package taskList;
+package taskList.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
