+++ src/main/java/taskList/tasks/ToDo.java
@@ -1,5 +1,4 @@
-package taskList;
-
+package taskList.tasks;
 
 public class ToDo extends Task {
