+++ text-ui-test/runtest.bat
@@ -8,6 +8,7 @@ if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
 javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\taskList\commands\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\taskList\tasks\*.java
 javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\taskList\*.java
 
 IF ERRORLEVEL 1 (
